requirements.txt
gradio
python-dotenv
Pillow